refactor: move custom components out of ui/ into shared/
components/ui/ is reserved for shadcn-generated primitives (.fallowrc.jsonc
ignores all of its exports as scaffolded "full kits"). The custom wrappers
that had accumulated there don't fit that contract, so relocate them to
components/shared/ alongside the existing DestructiveConfirmDialog:

- FilterPill, dialog-icon-header, stat-row, segmented-control (+ browser tests)
- Rewrite their relative ./button / ./dialog imports to the @/ alias,
  matching the shared/ convention
- Update all import sites, the Storybook primitives story, and DESIGN.md
- De-export StatRowTone / DialogIconHeaderTone: now outside the ui/
  ignoreExports zone, these internal-only types must not be public exports

ui/ now contains only shadcn primitives.
