+++ src/renderer/src/components/shared/FilterPill.browser.test.tsx
