+++ src/renderer/src/components/shared/segmented-control.browser.test.tsx
