+++ src/renderer/src/components/shared/segmented-control.tsx
