+++ src/renderer/src/components/sidebar/SyncConfirmDialog.tsx
@@ -1,6 +1,8 @@
 import { FolderSync, Loader2 } from 'lucide-react'
 import React, { useCallback } from 'react'
 
+import { DialogIconHeader } from '@/renderer/src/components/shared/dialog-icon-header'
+import { StatRow } from '@/renderer/src/components/shared/stat-row'
 import { Button } from '@/renderer/src/components/ui/button'
 import {
   Dialog,
@@ -9,8 +11,6 @@ import {
   DialogFooter,
   DialogHeader,
 } from '@/renderer/src/components/ui/dialog'
-import { DialogIconHeader } from '@/renderer/src/components/ui/dialog-icon-header'
-import { StatRow } from '@/renderer/src/components/ui/stat-row'
 import { useExecuteSync } from '@/renderer/src/hooks/useExecuteSync'
 import { shouldShowSyncConfirm } from '@/renderer/src/lib/syncHelpers'
 import { useAppDispatch, useAppSelector } from '@/renderer/src/redux/hooks'
